Update GitHub Actions Dependabot Groups (#84)
# Pull Request

## Description

This pull request includes a change to the `.github/dependabot.yml` file
to update the configuration for GitHub Actions dependencies. The most
important change is renaming the group and specifying the types of
updates to be applied.

Configuration updates:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L13-R18):
Renamed the `github-action-dependencies` group to `github-actions` and
added `update-types` to specify "patch" and "minor" updates.