Remove Actionlint Job (#94)
# Pull Request

## Description

This pull request includes updates to the YAML linting configuration and
the code quality workflow. The changes focus on improving the
configuration rules and removing redundant steps in the workflow.

Updates to YAML linting configuration:

*
[`.github/super-linter-configs/.yaml-lint.yml`](diffhunk://#diff-cdea8bea457c52a8c0651465a847dcbc52c6abd1fb370c531b89a3214efad8bbL7-R8):
Modified the `comments` rule to require a minimum of one space from the
content.

Simplification of code quality workflow:

*
[`.github/workflows/code-quality.yml`](diffhunk://#diff-22e4cd39aff0d5776e7be7d8a8457193836427c286230ca6836710497dcda0e8L49-L63):
Removed the `run-actionlint` job, which included steps for checking out
the repository, installing actionlint, and checking workflow files.