Add Code of Conduct (#98)
# Pull Request

## Description

This pull request introduces a new Code of Conduct document to the
project, outlining the expectations for community behavior and the
enforcement guidelines. The most important changes include the addition
of sections detailing the community's pledge, standards for behavior,
enforcement responsibilities, scope, and the process for reporting and
addressing violations.

Key additions:

*
[`docs/CODE_OF_CONDUCT.md`](diffhunk://#diff-d00bdf4d7740a49ab589446623c6c4e04fe86e52b92e3069da8ecbec24587005R1-R131):
Added a comprehensive Code of Conduct document adapted from the
Contributor Covenant, version 2.1, including sections on community
pledge, standards, enforcement responsibilities, scope, enforcement
process, and guidelines for consequences.