Scan GitHub Actions with CodeQL (#104)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
for running CodeQL analysis on the repository. The most important
changes include adding a new job definition for CodeQL analysis,
configuring the necessary permissions, and specifying the steps to
checkout the repository, initialize CodeQL, and perform the analysis.

Improvements to code quality analysis:

*
[`.github/workflows/code-quality.yml`](diffhunk://#diff-22e4cd39aff0d5776e7be7d8a8457193836427c286230ca6836710497dcda0e8R73-R91):
Added a new job named `run-codeql-analysis` to perform CodeQL analysis
with the necessary steps and permissions.