Clean GitHub Action Caches (#118)
# Pull Request

## Description

This pull request introduces a new GitHub Actions workflow to clean
caches on a scheduled basis. The workflow is designed to run on the
first day of each month and can also be triggered manually.

Key changes include:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dR1-R26):
Added a new workflow named "Clean Caches" that schedules a job to clean
GitHub Action caches monthly and provides a manual trigger option. The
job includes steps to check out the repository, authenticate GitHub CLI,
and list and delete caches.