Fix Cache Deletion Permissions and Command (#121)
# Pull Request

## Description

This pull request includes a significant update to the cache management
process in the GitHub Actions workflow. The changes simplify the cache
deletion process and update the environment variable for the GitHub
token.

Changes to cache management:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL21-R24):
Simplified the cache deletion process by replacing the manual API call
with the `gh cache delete --all` command.
*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL21-R24):
Updated the environment variable for the GitHub token from `GH_TOKEN` to
`secrets.GH_TOKEN`.