Update Dependency Review (#126)
# Pull Request

## Description

This pull request includes changes to the GitHub Actions workflow for
dependency review. The changes enhance the permissions and configuration
for the dependency review job.

Enhancements to dependency review workflow:

*
[`.github/workflows/dependency-review.yml`](diffhunk://#diff-7cdd3ccec44c8ba176bdc3b9ef54c3f56aa210a1a4e2bb5f79d87b1e50314a18R14-R15):
Added `pull-requests: write` permission to the `dependency-review` job.
*
[`.github/workflows/dependency-review.yml`](diffhunk://#diff-7cdd3ccec44c8ba176bdc3b9ef54c3f56aa210a1a4e2bb5f79d87b1e50314a18R24-R25):
Configured the dependency review action to comment a summary in the pull
request on failure.