Update Lefthook Configuration (#130)
# Pull Request

## Description

This pull request introduces a new validation step for `lefthook`
configuration and updates the `lefthook.yml` file to include this
validation. Additionally, it modifies the structure of the
`lefthook.yml` file to improve readability and maintainability.

New validation step:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR33-R40):
Added a new `lefthook-validate` target to validate the `lefthook`
configuration.

Updates to `lefthook.yml`:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR2-R15):
Added `min_version` and `colors` configuration options.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR2-R15):
Changed `jobs` to `commands` and updated the structure to include the
new `lefthook-validate` command.