Run Lefthook Validate Job (#131)
# Pull Request

## Description

This pull request introduces a new job in the GitHub Actions workflow to
validate code using Lefthook. The most important change is the addition
of the `lefthook-validate` job, which ensures that the code adheres to
predefined hooks and standards.

Changes to GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R92-R107):
Added a new job `lefthook-validate` to run Lefthook validation,
including steps to checkout the repository, install the latest version
of `uv`, and execute the Lefthook validation command.