Sync Repository Labels (#133)
# Pull Request

## Description

This pull request includes changes to streamline the GitHub repository's
label management and workflow processes. The most important changes are
the addition of a new labels configuration file, the removal of the
dependency review workflow, and the introduction of a label
synchronization workflow.

Label management improvements:

*
[`.github/other-configs/labels.yml`](diffhunk://#diff-2e7c23e626fa8c338aa942fce55a914e6e08b1643901fdd722f4168ea42c78fbR1-R76):
Added a comprehensive set of default GitHub labels, including categories
for issues, languages, components, and pull request sizes.

Workflow adjustments:

*
[`.github/workflows/dependency-review.yml`](diffhunk://#diff-7cdd3ccec44c8ba176bdc3b9ef54c3f56aa210a1a4e2bb5f79d87b1e50314a18L1-L25):
Removed the dependency review workflow to simplify the repository's
workflows.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6R1-R29):
Introduced a new workflow to automatically synchronize labels based on
the `labels.yml` configuration file. This workflow triggers on pushes to
the main branch and updates the repository's labels accordingly.