Further Update Lefthook Git Hooks (#138)
# Pull Request

## Description

This pull request includes changes to improve the formatting and output
of the pre-commit hooks and gitleaks detection. The most important
changes are the removal of the `-q` flag from the `just` commands to
ensure full output visibility and the addition of a summary output
configuration in `lefthook.yml`.

Improvements to pre-commit hooks and gitleaks detection:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL26-R31):
Removed the `-q` flag from the `just` commands to ensure full output
visibility for gitleaks detection and other pre-commit checks.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR5-R20):
Added a summary output configuration and removed the `-q` flag from the
`just` commands to ensure full output visibility for pre-commit hooks.