Remove pinact-verify Job (#174)
# Pull Request

## Description

This pull request simplifies the GitHub Actions workflow by removing an
unused job and its associated steps.

Workflow cleanup:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L128-L141):
Removed the `pinact-verify` job, which included steps for checking out
the repository and installing the latest version of `uv`. This job was
deemed unnecessary and has been eliminated to streamline the workflow.