Bump GitHub Action Versions (#175)
# Pull Request

## Description

This pull request updates several dependencies and configurations in the
repository's workflows and tools. The changes include upgrading action
versions in GitHub workflows, modifying branch configuration, and
updating the minimum required version of `lefthook`.

### Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L81-R87):
Updated the `setup-uv` action to version `6.0.1` and the `codeql-action`
actions (`upload-sarif`, `init`, and `analyze`) to version `v3.28.16`
for improved compatibility and features.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L81-R87)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L104-R102)
[[3]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L122-R123)
*
[`.github/workflows/generate-svg.yml`](diffhunk://#diff-158690ab8298e24898f32018d36e7725006eb11cd1c455790cfaad090285a112L10-R10):
Changed the `push` branch configuration from `"main"` to `["main"]` to
align with YAML array syntax.

### Tool configuration update:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the minimum required version of `lefthook` from `1.11.8` to
`1.11.12` to ensure compatibility with newer features or fixes.