Swap SuperLinter to Full Version (#189)
# Pull Request

## Description

This pull request includes updates to the configuration for Dependabot
and GitHub Actions workflows, specifically focusing on dependency
maintenance and the usage of the `super-linter` tool.

### Dependabot configuration updates:
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28R2-L3):
Removed a comment related to maintaining dependencies for GitHub
Actions.
*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L17-R17):
Updated the `exclude-patterns` in the `updates` section to explicitly
exclude `super-linter/super-linter` instead of the broader
`super-linter/*`.

### GitHub Actions workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L27-R27):
Updated the `uses` reference for `super-linter` to point directly to
`super-linter/super-linter` instead of `super-linter/super-linter/slim`.