Set Zizmor persona to auditor (#216)
# Pull Request

## Description

This pull request updates several reusable workflow references in GitHub
Actions configuration files to a newer version and modifies a command in
the `Justfile` to use a different persona for the `zizmor` tool.

### GitHub Actions Workflow Updates:
*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow to version `v2025.06.16.01` for cleaning
caches.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L42-R42):
Updated the reusable workflows for both code checks and CodeQL analysis
to version `v2025.06.16.01`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L42-R42)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L54-R54)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL15-R15):
Updated the reusable workflow for pull request tasks to version
`v2025.06.16.01`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L18-R18):
Updated the reusable workflow for syncing labels to version
`v2025.06.16.01`.

### Command Update in `Justfile`:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL47-R47):
Changed the `zizmor` command to use the `auditor` persona instead of the
`pedantic` persona.