Update Reusable Workflows (#227)
# Pull Request

## Description

This pull request updates the reusable workflow versions across multiple
GitHub Actions workflow files to ensure compatibility with the latest
changes and improvements. The updates replace references to an older
version (`v2025.06.29.03`) with a newer version (`v2025.07.06.04`) for
better functionality and security.

Workflow updates:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reusable workflow reference for cleaning caches to the
latest version.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L45-R45):
Updated reusable workflow references for code checks and CodeQL analysis
to the latest version.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L45-R45)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L57-R57)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL15-R15):
Updated the reusable workflow reference for common pull request tasks to
the latest version.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L19-R19):
Updated the reusable workflow reference for syncing labels to the latest
version.