Add actionlint pre-commit check (#232)
# Pull Request

## Description

This pull request introduces a new `actionlint-check` command to the
project and updates the `lefthook` configuration to support it. The
changes ensure that Actionlint checks are integrated into the pre-commit
hooks and that the minimum required version of `lefthook` is updated.

### Integration of Actionlint:

*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR49-R56):
Added a new `actionlint-check` command to run Actionlint checks as part
of the project's validation processes.
*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dR23-R24):
Integrated `actionlint-check` into the `pre-commit` hooks to ensure
Actionlint checks are run automatically before commits.

### Lefthook Configuration Update:

*
[`lefthook.yml`](diffhunk://#diff-ad6a01e589b8b1b214ca310dbb8d2e4314f6c612b921050c73c97455de43884dL2-R2):
Updated the `min_version` of `lefthook` from `1.12.1` to `1.12.2` to
ensure compatibility with the new configuration.